init hab command only on i.mx6ul
