+++ cmd/otp.go
@@ -8,7 +8,6 @@
 package cmd
 
 import (
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"regexp"
@@ -17,41 +16,12 @@ import (
 	"github.com/usbarmory/crucible/fusemap"
 	"github.com/usbarmory/crucible/otp"
 
-	"github.com/usbarmory/tamago-example/internal/hab"
 	"github.com/usbarmory/tamago-example/shell"
-	"github.com/usbarmory/tamago/soc/nxp/snvs"
 )
-
-const habWarning = `
-████████████████████████████████████████████████████████████████████████████████
-
-                 !!!           READ CAREFULLY           !!!
-
-This command activates secure boot (HAB) on your device SoC by permanent OTP
-fusing of the argument SRK hash.
-
-Fusing OTP's is an **irreversible** action that permanently fuses values on the
-device. This means that your device will be able to only execute firmware
-signed with the corresponding private keys after programming is completed.
-
-In other words your device will stop acting as a generic purpose device and
-will be converted to *exclusive use of your own signed firmware releases*.
-
-████████████████████████████████████████████████████████████████████████████████
-`
 
 var fuseMap *fusemap.FuseMap
 
 func init() {
-	shell.Add(shell.Cmd{
-		Name:    "hab",
-		Args:    1,
-		Pattern: regexp.MustCompile(`^hab ([[:xdigit:]]+)$`),
-		Syntax:  "<srk table hash>",
-		Help:    "HAB activation (use with extreme caution)",
-		Fn:      habCmd,
-	})
-
 	shell.Add(shell.Cmd{
 		Name:    "otp",
 		Args:    2,
@@ -85,32 +55,6 @@ func readOTP(bank int, word int) (res string, err error) {
 	}
 
 	return "", errors.New("invalid OTP register")
-}
-
-func habCmd(console *shell.Interface, arg []string) (res string, err error) {
-	srk, err := hex.DecodeString(arg[0])
-
-	if err != nil {
-		return "", fmt.Errorf("invalid SRK hash, %v", err)
-	}
-
-	fmt.Fprintf(console.Output, habWarning)
-
-	if !console.Confirm("Are you sure? (y/n) ") {
-		return "command cancelled", nil
-	}
-
-	if OCOTP == nil || SNVS == nil {
-		return "", errors.New("unavailable")
-	}
-
-	ssm := SNVS.Monitor()
-
-	if ssm.State != snvs.SSM_STATE_NONSECURE {
-		return "", fmt.Errorf("invalid state (%#.4b)", ssm.State)
-	}
-
-	return "", hab.Activate(OCOTP, SNVS, srk)
 }
 
 func otpCmd(_ *shell.Interface, arg []string) (res string, err error) {
